requirements.txt
Django==1.8.3
djangorestframework==3.1.3
mysqlclient==1.3.6